more warnings fixed and compiles on windows

+++ src/engine/inc/material.h
@@ -51,9 +51,9 @@ public:
 	const std::vector<Notation>& getPlacementsOfPiece(ChessPiece piece) const;
 
 	u32 getValue() const;
-	u32 getCount() const;
-	u32 getPieceCount(ChessPiece piece) const;
-    u32 getPieceCount(PieceType pieceType) const;
+	size_t getCount() const;
+	size_t getPieceCount(ChessPiece piece) const;
+	size_t getPieceCount(PieceType pieceType) const;
 	
 private:
 	typedef std::array<std::vector<Notation>, (size_t)PieceType::KING> MaterialGrid;
